huge refactor to make  calculating summary more accurate.
