+++ internal/smash/helper.go
